pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clarity"
version = "1.0.0"
description = "Clarity: AI DevOps Copilot — Intelligent incident management and code quality platform"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pandas>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "psutil>=5.9.0",
    "alembic>=1.13.0",
    "slowapi>=0.1.9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["clarity*"]

[project.optional-dependencies]
context = [
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.10.0",
    "sentence-transformers>=2.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
clarity = "clarity.cli.app:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
